requirements.txt
attrs==17.3.0
Automat==0.6.0
Babel==2.3.4
blinker==1.4
click==6.7
constantly==15.1.0
coverage==4.0.3
decorator==4.0.9
defusedxml==0.4.1
dominate==2.2.0
Flask==0.10.1
Flask-Babel==0.10.0
Flask-Bootstrap==3.3.5.7
Flask-HTTPAuth==3.1.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
httpauth==0.3
hyperlink==17.3.1
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.8
#lxml==3.4.4
MarkupSafe==0.23
pbr==1.9.1
Pillow==3.0.0
pyBarcode==0.8b1
pymssql==2.1.1
pypiwin32==219
python-docx==0.8.6
python3-openid==3.0.10
pytz==2016.4
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.14
#sqlalchemy-migrate==0.11.0
sqlparse==0.1.19
Tempita==0.5.2
#Twisted==17.9.0
virtualenv==15.0.1
visitor==0.1.3
Werkzeug==0.11.9
Whoosh==2.7.4
WTForms==2.1
xlwt==1.0.0
#zope.interface==4.4.3
